+++ tests/Integration/AbstractTechproductsTest.php
@@ -15,6 +15,7 @@ use Solarium\Core\Client\ClientInterface;
 use Solarium\Core\Client\Request;
 use Solarium\Exception\HttpException;
 use Solarium\Exception\RuntimeException;
+use Solarium\Exception\UnexpectedValueException;
 use Solarium\Plugin\BufferedAdd\Event\AddDocument as BufferedAddAddDocumentEvent;
 use Solarium\Plugin\BufferedAdd\Event\Events as BufferedAddEvents;
 use Solarium\Plugin\BufferedAdd\Event\PostCommit as BufferedAddPostCommitEvent;
@@ -591,6 +592,191 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertSame(32, $queryGroup->getMatches());
         $this->assertSame(0, $queryGroup->getNumFound());
         $this->assertNull($queryGroup->getMaximumScore());
+    }
+
+    public function testMoreLikeThisComponent()
+    {
+        $select = self::$client->createSelect();
+        $select->setQuery('apache');
+        $select->setSorts(['id' => SelectQuery::SORT_ASC]);
+
+        $moreLikeThis = $select->getMoreLikeThis();
+        $moreLikeThis->setFields('manu,cat');
+        $moreLikeThis->setMinimumDocumentFrequency(1);
+        $moreLikeThis->setMinimumTermFrequency(1);
+        $moreLikeThis->setInterestingTerms('details');
+
+        $result = self::$client->select($select);
+        $this->assertSame(2, $result->getNumFound());
+
+        $iterator = $result->getIterator();
+        $mlt = $result->getMoreLikeThis();
+
+        $document = $iterator->current();
+        $this->assertSame('SOLR1000', $document->id);
+        $mltResult = $mlt->getResult($document->id);
+        // actual max. score isn't consistent across Solr 7 and 8, server and cloud
+        // but it must always be a float
+        $this->assertIsFloat($mltResult->getMaximumScore());
+        $this->assertSame(1, $mltResult->getNumFound());
+        $mltDoc = $mltResult->getIterator()->current();
+        $this->assertSame('UTF8TEST', $mltDoc->id);
+
+        $iterator->next();
+        $document = $iterator->current();
+        $this->assertSame('UTF8TEST', $document->id);
+        $mltResult = $mlt->getResult($document->id);
+        $this->assertIsFloat($mltResult->getMaximumScore());
+        $this->assertSame(1, $mltResult->getNumFound());
+        $mltDoc = $mltResult->getIterator()->current();
+        $this->assertSame('SOLR1000', $mltDoc->id);
+
+        // Solr 7 doesn't support mlt.interestingTerms for MoreLikeThisComponent
+        // Solr 8: "To use this parameter with the search component, the query cannot be distributed."
+        // https://solr.apache.org/guide/morelikethis.html#common-handler-and-component-parameters
+        if (8 <= self::$solrVersion && $this instanceof AbstractServerTest) {
+            // with 'details', interesting terms are an associative array of terms and their boost values
+            $interestingTerms = $mlt->getInterestingTerm($document->id);
+            $this->assertSame('cat:search', key($interestingTerms));
+            $this->assertSame(1.0, current($interestingTerms));
+
+            $moreLikeThis->setInterestingTerms('list');
+            $result = self::$client->select($select);
+            $document = $result->getIterator()->current();
+            $mlt = $result->getMoreLikeThis();
+
+            // with 'list', interesting terms are a numeric array of strings
+            $interestingTerms = $mlt->getInterestingTerm($document->id);
+            $this->assertSame(0, key($interestingTerms));
+            $this->assertSame('cat:search', current($interestingTerms));
+
+            $moreLikeThis->setInterestingTerms('none');
+            $result = self::$client->select($select);
+            $document = $result->getIterator()->current();
+            $mlt = $result->getMoreLikeThis();
+
+            // with 'none', interesting terms aren't available for the MLT result
+            $this->expectException(UnexpectedValueException::class);
+            $this->expectExceptionMessage('interestingterms is none');
+            $mlt->getInterestingTerm($document->id);
+        }
+    }
+
+    /**
+     * There are a number of open issues that show MoreLikeThisHandler doesn't work as expected in SolrCloud mode.
+     *
+     * @see https://issues.apache.org/jira/browse/SOLR-4414
+     * @see https://issues.apache.org/jira/browse/SOLR-5480
+     *
+     * @group skip_for_solr_cloud
+     */
+    public function testMoreLikeThisQuery()
+    {
+        $query = self::$client->createMoreLikethis();
+
+        // the document we query to get similar documents for
+        $query->setQuery('id:SP2514N');
+        $query->setMltFields('manu,cat');
+        $query->setMinimumDocumentFrequency(1);
+        $query->setMinimumTermFrequency(1);
+        $query->setInterestingTerms('details');
+        // ensures we can consistently test for boost=1.0
+        $query->setBoost(false);
+        $query->setMatchInclude(true);
+        $query->createFilterQuery('stock')->setQuery('inStock:true');
+
+        $resultset = self::$client->moreLikeThis($query);
+        $this->assertSame(7, $resultset->getNumFound());
+
+        $iterator = $resultset->getIterator();
+        $document = $iterator->current();
+        $this->assertSame('6H500F0', $document->id);
+
+        $matchDocument = $resultset->getMatch();
+        $this->assertSame('SP2514N', $matchDocument->id);
+
+        // with 'details', interesting terms are an associative array of terms and their boost values
+        $interestingTerms = $resultset->getInterestingTerms();
+        $this->assertSame('cat:electronics', key($interestingTerms));
+        $this->assertSame(1.0, current($interestingTerms));
+
+        $query->setInterestingTerms('list');
+        $resultset = self::$client->moreLikeThis($query);
+
+        // with 'list', interesting terms are a numeric array of strings
+        $interestingTerms = $resultset->getInterestingTerms();
+        $this->assertSame(0, key($interestingTerms));
+        $this->assertSame('electronics', current($interestingTerms));
+
+        $query->setInterestingTerms('none');
+        $resultset = self::$client->moreLikeThis($query);
+
+        // with 'none', interesting terms aren't available for the result set
+        $this->expectException(UnexpectedValueException::class);
+        $this->expectExceptionMessage('interestingterms is none');
+        $resultset->getInterestingTerms();
+    }
+
+    /**
+     * There are a number of open issues that show MoreLikeThisHandler doesn't work as expected in SolrCloud mode.
+     *
+     * @see https://issues.apache.org/jira/browse/SOLR-4414
+     * @see https://issues.apache.org/jira/browse/SOLR-5480
+     *
+     * @group skip_for_solr_cloud
+     */
+    public function testMoreLikeThisStream()
+    {
+        $query = self::$client->createMoreLikethis();
+
+        // the supplied text we want similar documents for
+        $text = <<<EOT
+            Samsung SpinPoint P120 SP2514N - hard drive - 250 GB - ATA-133
+            7200RPM, 8MB cache, IDE Ultra ATA-133, NoiseGuard, SilentSeek technology, Fluid Dynamic Bearing (FDB) motor
+            EOT;
+
+        $query->setQuery($text);
+        $query->setQueryStream(true);
+        $query->setMltFields('name,features');
+        $query->setMinimumDocumentFrequency(1);
+        $query->setMinimumTermFrequency(1);
+        $query->setInterestingTerms('details');
+        // ensures we can consistently test for boost=1.0
+        $query->setBoost(false);
+        $query->setMatchInclude(true);
+        $query->createFilterQuery('stock')->setQuery('inStock:true');
+
+        $resultset = self::$client->moreLikeThis($query);
+        $this->assertSame(3, $resultset->getNumFound());
+
+        $iterator = $resultset->getIterator();
+        $document = $iterator->current();
+        $this->assertSame('SP2514N', $document->id);
+
+        // there is no match document to return, even with matchinclude true
+        $this->assertNull($resultset->getMatch());
+
+        // with 'details', interesting terms are an associative array of terms and their boost values
+        $interestingTerms = $resultset->getInterestingTerms();
+        // which term comes first differs between Solr 7 and 8, but it must always be a string
+        $this->assertIsString(key($interestingTerms));
+        $this->assertSame(1.0, current($interestingTerms));
+
+        $query->setInterestingTerms('list');
+        $resultset = self::$client->moreLikeThis($query);
+
+        // with 'list', interesting terms are a numeric array of strings
+        $interestingTerms = $resultset->getInterestingTerms();
+        $this->assertSame(0, key($interestingTerms));
+        $this->assertIsString(current($interestingTerms));
+
+        $query->setInterestingTerms('none');
+        $resultset = self::$client->moreLikeThis($query);
+
+        // with 'none', interesting terms aren't available for the result set
+        $this->expectException(UnexpectedValueException::class);
+        $this->expectExceptionMessage('interestingterms is none');
+        $resultset->getInterestingTerms();
     }
 
     public function testQueryElevation()
@@ -2053,7 +2239,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $this->assertSame('HTML Test Title', $document['title'][0], 'Written document does not contain extracted title');
         $this->assertMatchesRegularExpression('/^HTML Test Title\s+HTML Test Body$/', trim($document['content'][0]), 'Written document does not contain extracted result');
 
-        // now cleanup the document the have the initial index state
+        // now cleanup the document to have the initial index state
         $update = self::$client->createUpdate();
         $update->addDeleteQuery('cat:extract-test');
         $update->addCommit(true, true);